refactor server utility.
